Fix write length in snprinft calls
The remaining number of bytes that can we written to the input buffer for page content was flawed.
`offset` gets overwritten and does not properly track the remaining amount of bytes available in the buffer.
Now using the difference between the continuously advanced pointer `pos` and our initial buffer pointer `buf`
is now used instead.