Improve Html escaping logic
Previously we used consecutive calls to strncat to escape user provided Html content which led to disastrous performance
This lead to redundant reads of content by strncat ending up with an escape algorithm of O(n²) complexity.

We now use the same logic as in __html_create_content and advance a pointer instead.