Don't include title and dates in feed content
As lots of feed readers tend to render a pages title and creation/update
dates independently from the content we now separate plain content from it's
representation as a blog page.

Previous calls to strtok were not necessary for the simple printing we did
so we simply replaced it and now write the content in one go to the destination (file).

The actual content parsing logic is now extracted into it's own function
residing in page.c and works similar to how the already existing header parsing
function works.

This led to the appending of content to our content array now not happening inside
process_page but instead after it successfully created a (Html) page. Both a page's headers
and it's content are now returned and ownership transferred back to the caller of process_page.