Simplify handling of different page types
We refined our previous approach of handling the blog entry point and
mere posts differently.

That simplification now looks like the following:

Pages that are not to be listed as part of the list of pages shown in
the blog entry point now only get their page (post) creation and
modification dates displayed at the bottom. Title and the reference to
the blog entry point are omitted on purpose.

Every other kind of page that is to be considered a proper post has
it's title as well as said back reference link dynamically inserted.
The provision of pages to be omitted now is possible through a new
_SITE_EXT_EXEMPT_EXTRA make variable. The escaping issues here are
horrendous but tolerable for now.

The two notable exemptions here are the site index (newly introduced
through the _SITE_EXT_SITE_INDEX make variable) and the blog entry
point (now renamed to _SITE_EXT_BLOG_INDEX). Both of these exceptions
are omitted through adding them as defaults of the _SITE_EXCEMPT_LIST.